Feature/gina milestone2 (#17)
* Updated report to reflect task 1 changes

In task 1, I created a new dev branch for .gitignore changes, added a workflow on how to use this branch, mentioned rebase commands and updated the .gitignore file to include large dataset files, ML model files, media/assets and OS files.

* Report table of contents added to reflect progress

Placeholder titles added to table of contents based on expected sections in milestone 2 tasks. #12 changed to reflect changes made in document.

* Update Task 2

Added task 2 questions and answers to the report.md. Updated corresponding table of contents

* Updated task 5

The report was updated with the information from Task 5. The report was reformatted to reflect the added information and improve readability

* Updated report with final code testing

Completed report for milestone 2 with final code testing and notes about tag and release.

* Updated README

Updated project structure and workflow to reflect processes completed in milestone 2